[ft #[163518673]] add user endpoint
